Update import script to pull from guidebook dump
Using actual data when we're developing the app will give us a better
idea of what data the users will be seeing.

Also, importing the data during a database seed will save the user from
having to worry about importing data – it will already be imported
before they start using the app.

There are some fields that are imported from the guidebook that we don't
yet have database fields for:

- Fax
- Miscellaneous
- What to Bring
- Eligible Population
- Faith Based

As we add these fields, we should update the database seed file to
match.