Add human service categories from OpenEligibility
Problem:

The `services` field for organizations is wildly inconsistent.
Eventually, we'd like all organizations to be structured under a
standard taxonomy to make browsing and searching more managable.

The best example we've found of a community-accepted taxonomy is the
OpenEligibility taxonomy (http://about.auntbertha.com/openeligibility).

Solution:

We'd like each organization to belong to *at least* one of the
OpenEligibility categories. To make sure this happens, we've:

- Seeded the database with all of the leaf nodes for the OpenEligibility
  taxonomy
- Improved the form field for assigning organizations to a category
  - Add Selectize for services form fields
  - See http://brianreavis.github.io/selectize.js/
- Added a conditional validation to make sure each organization has at
  least one taxonomy category
  - Organizations that don't pass this validation will be displayed with
    a red flag on the index page.