Remove importing
Problem:

After getting a better sense of APD's use cases,
it's likely that we'll only run the import script once.
We've already moved the import script to run when we seed the database,
so there's no need to have additional import functionality in the UI.

Solution:

Remove the import table from the database,
as well as all references to it from the UI and code base.