Remove service tags and taxonomy
Problem:

The services and taxonomy account for most of the effort we've been
putting into this app so far. As LinkSF has demonstrated, they're not an
essential part of an MVP, so it makes sense to scrap them in favor of
four or five broad base-level categories.

Solution:

Remove all service-related code.