Add miscellaneous field to organizations
Problem:

When information about an organization is not parsed correctly,
the info that could not be detected is returned by the import script as
a catch-all 'miscellaneous' field. Since there was no corresponding
field in the database, the information that could not be parsed was
lost.

Solution:

Add a miscellaneous field to store the information that could not be
captured.

Next steps:

Go through organizations that have data in the miscellaneous field to
further improve the import script